+++ src/Controller/UserController.php
@@ -8,12 +8,19 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Security;
 
 /**
  * @Route("/user")
  */
 class UserController extends AbstractController
 {
+      private $security;
+
+      public function __construct(Security $security)
+      {
+          $this->security = $security;
+      }
     /**
      * @Route("/", name="user_index", methods={"GET"})
      */
@@ -33,6 +40,10 @@ class UserController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        if (!$this->security->isGranted('ROLE_ADMIN')) {
+            return $this->render('default/noaccess.html.twig');
+        }
+
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
@@ -56,6 +67,9 @@ class UserController extends AbstractController
      */
     public function show(User $user): Response
     {
+        if (!$this->security->isGranted('ROLE_ADMIN')) {
+            return $this->render('default/noaccess.html.twig');
+        }
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
@@ -66,6 +80,10 @@ class UserController extends AbstractController
      */
     public function edit(Request $request, User $user): Response
     {
+
+        if (!$this->security->isGranted('ROLE_ADMIN')) {
+            return $this->render('default/noaccess.html.twig');
+        }
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -88,6 +106,9 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
+        if (!$this->security->isGranted('ROLE_ADMIN')) {
+            return $this->render('default/noaccess.html.twig');
+        }
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($user);
